initial split by llm
